# docs/swagger.yaml
definitions:
...
  open_req.DeptQueryReq:
    properties:
      likeName:
        type: string
    type: object
...
  open_req.IdsReq:
    properties:
      ids:
        description: Ids
        items:
          type: integer
        type: array
    type: object
...
  open_resp.DeptInfoResp:
    properties:
      code:
        description: Code
        type: string
      id:
        description: Id
        type: integer
      isHide:
        description: IsHide
        type: integer
      name:
        description: Name
        type: string
      orgId:
        description: OrgId
        type: integer
      outOrgDepartmentCode:
        description: OutOrgDepartmentCode
        type: string
      outOrgDepartmentId:
        description: OutOrgDepartmentId
        type: string
      outOrgDepartmentParentId:
        description: OutOrgDepartmentParentId
        type: string
      parentId:
        description: ParentId
        type: integer
      path:
        description: Path
        type: string
      sort:
        description: Sort
        type: integer
      status:
        description: Status
        type: integer
    type: object
  open_resp.DeptMemberListResp:
    properties:
      code:
        description: Code
        type: string
      id:
        description: Id
        type: integer
      isHide:
        description: IsHide
        type: integer
      name:
        description: Name
        type: string
      orgId:
        description: OrgId
        type: integer
      outOrgDepartmentCode:
        description: OutOrgDepartmentCode
        type: string
      outOrgDepartmentId:
        description: OutOrgDepartmentId
        type: string
      outOrgDepartmentParentId:
        description: OutOrgDepartmentParentId
        type: string
      parentId:
        description: ParentId
        type: integer
      path:
        description: Path
        type: string
      sort:
        description: Sort
        type: integer
      status:
        description: Status
        type: integer
      userList:
        items:
          $ref: '#/definitions/open_resp.SimpleMemberInfo'
        type: array
    type: object
...
  open_resp.RoleInfoResp:
    properties:
      id:
        type: integer
      name:
        type: string
      orgId:
        type: integer
      roleGroupId:
        type: integer
    type: object
  open_resp.SimpleMemberInfo:
    properties:
      nickname:
        type: string
      status:
        description: Status
        type: integer
      userId:
        type: integer
    type: object
...
  req.CreateManageGroup:
    properties:
      groupType:
        description: 类型 1.系统管理组 2.普通管理组 3.bjx-普通管理员 4.bjx-普通成员,5.bjx-用户创建
        type: integer
      name:
        description: 名称
        type: string
      optAuth:
        description: 权限项
        items:
          type: string
        type: array
    type: object
...
  req.DeleteManageGroup:
    properties:
      id:
        description: ID
        type: integer
    type: object
...
  req.IdsReq:
    properties:
      ids:
        description: Ids ID列表
        items:
          type: integer
        type: array
    type: object
...
  req.UpdateManageGroup:
    properties:
      id:
        description: ID
        type: integer
      name:
        description: 名称
        type: string
    type: object
  req.UpdateManageGroupContents:
    properties:
      authToken:
        description: AuthToken 校验的 token，做一些特殊操作时，需要带上验证 token，如：更换管理员。
        type: string
      id:
        description: Id 管理组ID
        type: integer
      key:
        description: Key 属性名称 user_ids|app_package_ids|dept_ids|role_ids|app_ids
        type: string
      sourceFrom:
        description: SourceFrom 调用来源。默认为空，如果是极星，则传 `polaris`
        type: string
      valueIf:
        description: 由于 Values 值是特定的 []int64，为了兼容已经存在的接口，再加一个字段 valueIf（value interface），用于存放其他值。这两者互斥，如果
          values 有值，则优先使用 values。
        type: object
      values:
        description: Values ID列表
        items:
          type: integer
        type: array
    type: object
...
  resp.CacheUserInfoBo:
    properties:
      corpId:
        type: string
      orgId:
        type: integer
      outUserId:
        type: string
      sourceChannel:
        type: string
      userId:
        type: integer
    type: object
...
  resp.GetOperationConfigResp:
    properties:
      optAuthExtraInfo:
        additionalProperties: true
        description: OptAuthExtraInfo 权限项的额外信息
        type: object
      optAuthList:
        description: OptAuthList 权限项列表
        items:
          $ref: '#/definitions/resp.GetOperationConfigRespItem'
        type: array
    type: object
  resp.GetOperationConfigRespItem:
    properties:
      code:
        description: Code 权限项标识
        type: string
      group:
        description: Group 权限组名称
        type: string
      groupCode:
        description: GroupCode 权限组标识
        type: string
      name:
        description: Name 权限项名称
        type: string
      showStatus:
        description: ShowStatus 权限项是否展示的状态。1展示；2不展示。
        type: integer
    type: object
...
  resp.ManageGroupDetailResp:
    properties:
      adminGroup:
        $ref: '#/definitions/resp.ManageGroupInfo'
        description: AdminGroup 管理组信息
      headData:
        $ref: '#/definitions/resp.ManageGroupHead'
        description: HeadData 附属信息
    type: object
  resp.ManageGroupHead:
    properties:
      appPackages:
        description: AppPackages
        items:
          $ref: '#/definitions/resp.ManageGroupHeadData'
        type: array
      apps:
        description: Apps
        items:
          $ref: '#/definitions/resp.ManageGroupHeadData'
        type: array
      depts:
        description: Depts
        items:
          $ref: '#/definitions/resp.ManageGroupHeadData'
        type: array
      roles:
        description: Roles
        items:
          $ref: '#/definitions/resp.ManageGroupHeadData'
        type: array
      users:
        description: Users
        items:
          $ref: '#/definitions/resp.ManageGroupHeadDataUser'
        type: array
    type: object
  resp.ManageGroupHeadData:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  resp.ManageGroupHeadDataUser:
    properties:
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
    type: object
  resp.ManageGroupInfo:
    properties:
      appIds:
        description: 应用权限
        items:
          type: integer
        type: array
      appPackageIds:
        description: 应用包权限
        items:
          type: integer
        type: array
      createTime:
        type: string
      creator:
        type: integer
      deptIds:
        description: 部门权限
        items:
          type: integer
        type: array
      id:
        type: integer
      langCode:
        type: string
      name:
        type: string
      optAuth:
        description: 功能权限
        items:
          type: string
        type: array
      orgId:
        type: integer
      roleIds:
        description: 角色权限
        items:
          type: integer
        type: array
      type:
        description: 1 系统管理组 2普通管理组
        type: integer
      usageIds:
        description: 功能权限  【废弃】
        items:
          type: integer
        type: array
      userIds:
        description: 成员
        items:
          type: integer
        type: array
    type: object
  resp.ManageGroupTreeResp:
    properties:
      generalGroups:
        description: 普通管理组列表
        items:
          $ref: '#/definitions/resp.SimpleManageGroupInfo'
        type: array
      sysGroup:
        $ref: '#/definitions/resp.SimpleManageGroupInfo'
        description: 系统管理组
    type: object
...
  resp.SimpleManageGroupInfo:
    properties:
      id:
        type: integer
      langCode:
        type: string
      name:
        type: string
      optAuth:
        items:
          type: string
        type: array
      orgId:
        type: integer
    type: object
...
info:
  contact: {}
  title: Ying Sou Sou Apis
paths:
...
  /open/usercenter/api/v1/dept/children:
    post:
      consumes:
      - application/json
      description: 根据上级ID列表获取部门信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/open_req.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/open_resp.DeptInfoResp'
            type: array
        "400":
          description: ""
      summary: 根据上级ID列表获取部门信息列表
      tags:
      - 部门（OpenApi调用）
  /open/usercenter/api/v1/dept/list:
    post:
      consumes:
      - application/json
      description: 获取部门信息列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/open_resp.DeptInfoResp'
            type: array
        "400":
          description: ""
      summary: 获取部门信息列表
      tags:
      - 部门（OpenApi调用）
  /open/usercenter/api/v1/dept/list-by-ids:
    post:
      consumes:
      - application/json
      description: 根据部门ID列表获取部门信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/open_req.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/open_resp.DeptInfoResp'
            type: array
        "400":
          description: ""
      summary: 根据部门ID列表获取部门信息列表
      tags:
      - 部门（OpenApi调用）
  /open/usercenter/api/v1/dept/list-by-query-cond:
    post:
      consumes:
      - application/json
      description: 获取部门信息列表,根据条件
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/open_req.DeptQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/open_resp.DeptInfoResp'
            type: array
        "400":
          description: ""
      summary: 获取部门信息列表，根据条件
      tags:
      - 部门（OpenApi调用）
  /open/usercenter/api/v1/dept/list-have-member:
    post:
      consumes:
      - application/json
      description: 获取部门信息列表,包含成员
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/open_resp.DeptMemberListResp'
            type: array
        "400":
          description: ""
      summary: 获取部门信息列表,包含成员
      tags:
      - 部门（OpenApi调用）
...
  /open/usercenter/api/v1/role/list:
    post:
      consumes:
      - application/json
      description: 获取角色信息列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/open_resp.RoleInfoResp'
            type: array
        "400":
          description: ""
      summary: 获取角色信息列表
      tags:
      - 角色（OpenApi调用）
  /open/usercenter/api/v1/role/list-by-ids:
    post:
      consumes:
      - application/json
      description: 根据角色ID列表获取角色信息列表
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/open_resp.RoleInfoResp'
            type: array
        "400":
          description: ""
      summary: 根据角色ID列表获取角色信息列表
      tags:
      - 角色（OpenApi调用）
  /usercenter/api/v1/adminGroup/create:
    post:
      consumes:
      - application/json
      description: 管理组新增接口
      parameters:
      - description: 入参
        in: body
        name: .+
        required: true
        schema:
          $ref: '#/definitions/req.CreateManageGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: ""
      security:
      - token: []
      summary: 管理组新增
      tags:
      - 管理员
  /usercenter/api/v1/adminGroup/delete:
    delete:
      consumes:
      - application/json
      description: 管理组新增接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.DeleteManageGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: ""
      security:
      - token: []
      summary: 管理组删除
      tags:
      - 管理员
  /usercenter/api/v1/adminGroup/detail:
    get:
      consumes:
      - application/json
      description: 管理详情接口
      parameters:
      - description: 管理组ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ManageGroupDetailResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 管理详情
      tags:
      - 管理员
  /usercenter/api/v1/adminGroup/operationConfigs:
    get:
      consumes:
      - application/json
      description: 管理组权限项配置信息接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.GetOperationConfigResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 管理组权限项配置信息
      tags:
      - 管理员
  /usercenter/api/v1/adminGroup/tree:
    get:
      consumes:
      - application/json
      description: 管理组树状接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ManageGroupTreeResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 管理组树状
      tags:
      - 管理员
  /usercenter/api/v1/adminGroup/update:
    put:
      consumes:
      - application/json
      description: 管理组新增接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.UpdateManageGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: ""
      security:
      - token: []
      summary: 管理组修改
      tags:
      - 管理员
  /usercenter/api/v1/adminGroup/updateContents:
    put:
      consumes:
      - application/json
      description: 修改管理组包含内容接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.UpdateManageGroupContents'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: ""
      security:
      - token: []
      summary: 修改管理组包含内容
      tags:
      - 管理员
...
  /usercenter/inner/api/v1/auth/api-key-auth:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 校验ApiKey接口
      parameters:
      - description: apiKey
        in: query
        name: apiKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.CacheUserInfoBo'
        "400":
          description: ""
      security:
      - token: []
      summary: 校验ApiKey
      tags:
      - 组织
...
swagger: "2.0"